Fix slight change that affected all md5s

+++ Sources/AudioKit/Internals/Table/AKTable.swift
@@ -177,7 +177,7 @@ public class AKTable: NSObject, MutableCollection, Codable {
     public var phaseOffset: Int {
         @inline(__always)
         get {
-            Int(phase * Float(count))
+            Int(phase * Double(count))
         }
     }
